+++ lib/add_to_cart_animated_button.dart
@@ -1,1 +1,3 @@
 export 'add_to_cart.dart';
+export 'add_to_cart_controller.dart';
+export 'add_to_cart_theme.dart';